requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.4
django-celery-beat==2.1.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-jazzmin==2.4.1
django-pandas==0.6.2
django-rest-multiple-models==2.1.3
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
ifilter==0.2
kombu==5.0.2
numpy==1.19.5
panda==0.3.1
pandas==1.2.0
prompt-toolkit==3.0.10
PyJWT==1.6.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
